opt/requirements.txt
discord.py
supervisor
requests
